restructure repo to allow refactoring of analyzers
